requirements.txt
fastapi
uvicorn
pymupdf
pandas
nltk
regex
py2neo
sqlalchemy
sentence-transformers
qdrant-client
langchain
langgraph
networkx
pyvis
phonenumbers
tldextract
streamlit
faiss-cpu==1.7.4
numpy<2.0